+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Release 4.0.0:
  - Get rid of class `Issuer.IssuedCredentialResult`, replacing it with `KmmResult<Issuer.IssuedCredential>`
  - Add return types to function calls to `SubjectCredentialStore`
  - Change from list to single credential in parameter for `Holder.storeCredentials()`, changing name to `storeCredential()`
+ - Add `AuthenticationResponseFactory`: Builds an authentication response from request and response parameters
 
 Release 3.8.0:
  - Kotlin 2.0.0